Add support for systems in FIPS mode
When a system is in FIPS mode not all hashing algorithms are usable by default.
The most notable of these is MD5. When trying to generate an MD5 hash in FIPS
mode a ValueError exception is raised. Support was added to get around this for
uses such as this project by passing a "usedforsecurity=False" keyword
argument to the constructor. However, this is not universallty supported and on
some systems this constructor form will result in a TypeError as a result of an
unsuported keyword argument. I observed this on both my MacBook and on an
Ubuntu WSL instance when testing. A try/except block allows us to support both
configurations (hopefully).