Refactor for standalone mode and add standalone mode testing
Testing against the argument parsing was requested and a small refactor was the
only way I saw feasible to do so. This allows the main() function in
ioc_scanner.py to only perform command line argument parsing before calling the
core logic function. This also allowed me to add the pyfakefs package to make
testing this new functionality feasible.