DRY testing and add new hash to testblob.txt
Put the commonly used values like the test file for --file and repeatedly used
hash values as globals. Added the tests/targets/lorem.txt file's SHA256 to the
testblob.txt file. Tests for providing a file as a source for hashes will now
check that lorem.txt and its SHA256 are in the results. This makes sure we can
confirm that it is only detected if we provide a file.