Remove pyfakefs workaround for missing /dev/null
The apb requested GH Actions run for this repo failed on 2020-12-25. Upon
investigation I saw that the error was

FileExistsError: [Errno 17] File exists in the fake filesystem: '/dev/null'

I then checked for recent releases for pyfakefs (and pytest), and saw that
v4.3.3 came out on 2020-12-20. Looking into changes I saw that pyfakefs fixed
the null device name in
https://github.com/jmcgeheeiv/pyfakefs/commit/f1e5163665e900bff09c21a86f6e05e547c1b3cf
, which appears to have been the underlying problem that caused us to manually
create '/dev/null' in the first place to workaround pytest needing it for the
caplog fixture.